refactor: split pkg/client/client.go into daemon + http_poller + grpc_streamer
Follow-up B from the main-refactor branch comparison (task #5).

The 512-line pkg/client/client.go did four things at once: daemon
construction, watcher fan-out, HTTP poll loop, and the gRPC failover
state machine. Splitting along those seams makes each file a single
responsibility and shrinks the cognitive load when working on any one
of them.

Layout
------

- pkg/client/daemon.go (new) — CertDXClientDaemon struct + Make,
  certData, watchingCert, watchUpdate, ClientInit,
  AddCertToWatch{,Opt}, startWatchers, LoadConfigurationAndValidate{,Opt},
  Stop, GetCertificate.
- pkg/client/http_poller.go (new) — httpRequestCert, httpPollingCert,
  HttpMain.
- pkg/client/grpc_streamer.go (new) — GRPC_CLIENT_STATE constants,
  the new grpcStreamer struct that encapsulates the failover state
  machine, handleMain / handleFailover / handleFallback / handleRestart
  methods, and GRPCMain that drives one run.
- pkg/client/client.go — removed.

Behaviour
---------

Pure code-organisation slice. Every observable behaviour is preserved:

- The five state names (GRPC_STATE_*) and their transitions are
  unchanged. The dispatch loop in grpcStreamer.run mirrors the
  closure-based loop that lived inside GRPCMain before.
- Magic numbers in the gRPC retry path are now named constants
  (grpcRetryWindow = 5*time.Minute, grpcRetryBackoff = 15*time.Second).
  Same values, just labelled.
- Encapsulating the state machine in grpcStreamer instead of trailing
  closures over locals also drops the "race-free per-state cancel"
  shape introduced in slice 6: the dispatcher is still the only writer
  of sessionCtx / sessionCancel, the fallback only invokes the
  captured cancel.

Verification
------------

- go build ./... ✅
- go vet ./... ✅
- full e2e suite under test/e2e/ passes with -race (~256s).
